+++ src/pbfbench/abc/tool/shell.py
@@ -8,6 +8,7 @@ from typing import TYPE_CHECKING
 import pbfbench.abc.tool.config as tool_cfg
 import pbfbench.abc.topic.results.items as abc_topic_results
 import pbfbench.experiment.file_system as exp_fs
+import pbfbench.samples.shell as smp_sh
 
 if TYPE_CHECKING:
     from collections.abc import Iterator
@@ -16,13 +17,29 @@ if TYPE_CHECKING:
 class ArgBashLinesBuilder[R: abc_topic_results.Result](ABC):
     """Argument bash lines builder."""
 
-    def __init__(self, tool_data_result: R) -> None:
+    def __init__(
+        self,
+        tool_data_result: R,
+        working_exp_fs_manager: exp_fs.Manager,
+    ) -> None:
         """Initialize."""
-        self.__tool_data_result = tool_data_result
+        self._tool_data_result = tool_data_result
+        self._working_exp_fs_manager = working_exp_fs_manager
+        self._sample_sh_var_fs_manager = smp_sh.sample_sh_var_fs_manager(
+            self._working_exp_fs_manager,
+        )
 
     def tool_data_result(self) -> R:
         """Get tool data result."""
-        return self.__tool_data_result
+        return self._tool_data_result
+
+    def working_exp_fs_manager(self) -> exp_fs.Manager:
+        """Get working experiment file system manager."""
+        return self._working_exp_fs_manager
+
+    def sample_sh_var_fs_manager(self) -> smp_sh.smp_fs.Manager:
+        """Get sample shell variable file system manager."""
+        return self._sample_sh_var_fs_manager
 
     @abstractmethod
     def init_lines(self) -> Iterator[str]:
@@ -50,26 +67,26 @@ class Commands[N: tool_cfg.Names, O: tool_cfg.Options](ABC):
         working_exp_fs_manager: exp_fs.Manager,
     ) -> None:
         """Initialize."""
-        self.__arg_names_with_sh_lines_builders = arg_names_with_checked_inputs
-        self.__options = options
-        self.__working_exp_fs_manager = working_exp_fs_manager
+        self._arg_names_with_sh_lines_builders = arg_names_with_checked_inputs
+        self._options = options
+        self._working_exp_fs_manager = working_exp_fs_manager
 
     def options(self) -> O:
         """Get options."""
-        return self.__options
+        return self._options
 
     def working_exp_fs_manager(self) -> exp_fs.Manager:
         """Get working experiment file system manager."""
-        return self.__working_exp_fs_manager
+        return self._working_exp_fs_manager
 
     def commands(self) -> Iterator[str]:
         """Iterate over the tool commands."""
-        for result_lines_builder in self.__arg_names_with_sh_lines_builders.values():
+        for result_lines_builder in self._arg_names_with_sh_lines_builders.values():
             yield from result_lines_builder.init_lines()
         yield ("")
         yield from self.core_commands()
         yield ("")
-        for result_lines_builder in self.__arg_names_with_sh_lines_builders.values():
+        for result_lines_builder in self._arg_names_with_sh_lines_builders.values():
             yield from result_lines_builder.close_lines()
 
     @abstractmethod
@@ -79,7 +96,7 @@ class Commands[N: tool_cfg.Names, O: tool_cfg.Options](ABC):
 
     def lines_builder(self, name: N) -> ArgBashLinesBuilder:
         """Get result."""
-        return self.__arg_names_with_sh_lines_builders[name]
+        return self._arg_names_with_sh_lines_builders[name]
 
     def argument(self, name: N) -> str:
         """Get result."""