+++ src/pbfbench/experiment/run.py
@@ -4,8 +4,11 @@ from __future__ import annotations
 
 import logging
 import shutil
+import time
 from enum import StrEnum
 from typing import TYPE_CHECKING
+
+import rich.progress as rich_prog
 
 import pbfbench.abc.tool.config as abc_tool_cfg
 import pbfbench.abc.tool.visitor as abc_tool_visitor
@@ -17,7 +20,7 @@ import pbfbench.experiment.file_system as exp_fs
 import pbfbench.experiment.shell as exp_shell
 import pbfbench.samples.file_system as smp_fs
 import pbfbench.samples.status as smp_status
-from pbfbench import slurm, subprocess_lib
+from pbfbench import root_logging, slurm, subprocess_lib
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
@@ -31,17 +34,23 @@ class RunStats:
     """Experiment run stats."""
 
     @classmethod
-    def new(cls) -> RunStats:
+    def new(cls, data_exp_fs_manager: exp_fs.Manager) -> RunStats:
         """Create new run stats."""
-        return cls(0, None, None)
+        with smp_fs.TSVReader.open(
+            smp_fs.samples_tsv(data_exp_fs_manager.root_dir()),
+        ) as smp_tsv_in:
+            number_of_samples = sum(1 for _ in smp_tsv_in.iter_row_numbered_items())
+        return cls(number_of_samples, 0, None, None)
 
     def __init__(
         self,
+        number_of_samples: int,
         number_of_samples_to_run: int,
         samples_with_missing_inputs: Iterable[str] | None,
         samples_with_errors: Iterable[str] | None,
     ) -> None:
         """Init run stats."""
+        self.__number_of_samples = number_of_samples
         self.__number_of_samples_to_run = number_of_samples_to_run
         self.__samples_with_missing_inputs = (
             list(samples_with_missing_inputs)
@@ -51,6 +60,10 @@ class RunStats:
         self.__samples_with_errors = (
             list(samples_with_errors) if samples_with_errors is not None else []
         )
+
+    def number_of_samples(self) -> int:
+        """Get number of samples."""
+        return self.__number_of_samples
 
     def number_of_samples_to_run(self) -> int:
         """Get number of samples to run."""
@@ -72,7 +85,10 @@ class RunStats:
 class ErrorStatus(StrEnum):
     """Experiment error status."""
 
+    NO_WRITE_ACCESS = "no_write_access"
+    NO_READ_ACCESS = "no_read_access"
     NO_TOOL_ENV_WRAPPER_SCRIPT = "no_tool_env_wrapper_script"
+    WRONG_EXPERIMENT_CONFIG_SYNTAX = "wrong_experiment_config_syntax"
     DIFFERENT_EXPERIMENT = "different_experiment"
 
 
@@ -86,7 +102,16 @@ def run_experiment_on_samples(
     tool_connector: abc_tool_visitor.Connector,
 ) -> Status:
     """Run the experiment."""
-    exp_config = tool_connector.read_config(exp_config_yaml)
+    rw_check_result = _check_read_write_access(data_dir, working_dir)
+    if rw_check_result is not None:
+        return rw_check_result
+
+    try:
+        exp_config = tool_connector.read_config(exp_config_yaml)
+    except Exception:  # noqa: BLE001
+        return ErrorStatus.WRONG_EXPERIMENT_CONFIG_SYNTAX
+
+    _log_config(exp_config, tool_connector)
 
     match preparation_result := _prepare_experiment_file_systems(
         data_dir,
@@ -99,9 +124,11 @@ def run_experiment_on_samples(
         case _:
             data_exp_fs_manager, working_exp_fs_manager = preparation_result
 
-    run_stats = RunStats.new()
+    run_stats = RunStats.new(data_exp_fs_manager)
 
     samples_to_run = _get_samples_to_run(data_exp_fs_manager, run_stats)
+
+    _prepare_sample_directories(samples_to_run, working_exp_fs_manager)
 
     checked_inputs_samples_to_run, samples_with_missing_inputs, tool_inputs = (
         _filter_missing_inputs(
@@ -112,11 +139,16 @@ def run_experiment_on_samples(
             working_exp_fs_manager,
         )
     )
+
     _write_experiment_missing_inputs(
         samples_with_missing_inputs,
         working_exp_fs_manager,
         run_stats,
     )
+
+    if not checked_inputs_samples_to_run:
+        _LOGGER.info("No samples to run")
+        return run_stats
 
     _create_and_run_sbatch_script(
         tool_connector,
@@ -126,6 +158,8 @@ def run_experiment_on_samples(
         data_exp_fs_manager,
         working_exp_fs_manager,
     )
+
+    _wait_all_job_finish(checked_inputs_samples_to_run, working_exp_fs_manager)
 
     _write_experiment_errors(
         checked_inputs_samples_to_run,
@@ -144,7 +178,57 @@ def run_experiment_on_samples(
         samples_to_run,
     )
 
+    if run_stats.samples_with_missing_inputs() or run_stats.samples_with_errors():
+        _LOGGER.info(
+            "The list of samples with missing inputs"
+            " or which exit with errors is written to file: %s",
+            data_exp_fs_manager.errors_tsv(),
+        )
+
     return run_stats
+
+
+def _check_read_write_access(data_dir: Path, working_dir: Path) -> None | ErrorStatus:
+    """Check read and write access."""
+    #
+    # Check read/write access to data_dir and working_dir
+    #
+    try:
+        working_dir.mkdir(parents=True, exist_ok=True)
+    except OSError:
+        _LOGGER.critical("No write access to %s", working_dir)
+        return ErrorStatus.NO_WRITE_ACCESS
+
+    for test_dir in (data_dir, working_dir):
+        file_test = test_dir / "test_read_write.txt"
+        try:
+            file_test.write_text("test")
+        except OSError:
+            _LOGGER.critical("No write access to %s", test_dir)
+            return ErrorStatus.NO_WRITE_ACCESS
+        try:
+            file_test.read_text()
+        except OSError:
+            _LOGGER.critical("No read access to %s", test_dir)
+            file_test.unlink()
+            return ErrorStatus.NO_READ_ACCESS
+        file_test.unlink()
+
+    return None
+
+
+def _log_config(
+    exp_config: exp_cfg.Config,
+    tool_connector: abc_tool_visitor.Connector,
+) -> None:
+    """Log config."""
+    _LOGGER.info(
+        "Running experiment %s with tool %s for the topic %s.",
+        exp_config.name(),
+        tool_connector.tool_description().name(),
+        tool_connector.tool_description().topic().name(),
+    )
+    _LOGGER.debug("Experiment config:\n%s", exp_config.to_yaml_dump())
 
 
 def _prepare_experiment_file_systems[C: exp_cfg.Config](
@@ -162,19 +246,16 @@ def _prepare_experiment_file_systems[C: exp_cfg.Config](
     )
 
     if not data_exp_fs_manager.tool_env_script_sh().exists():
-        _LOGGER.critical(
-            "The experiment in the data directory"
-            " does not have the tool environment wrapper script.",
-        )
         return ErrorStatus.NO_TOOL_ENV_WRAPPER_SCRIPT
 
     if exp_checks.exists(data_exp_fs_manager):
-        if not exp_checks.is_same_experiment(data_exp_fs_manager, exp_config):
-            _LOGGER.critical(
-                "The experiment in the data directory"
-                " does not have the same configuration of the current experiment.",
-            )
-            return ErrorStatus.DIFFERENT_EXPERIMENT
+        same_exp_result = exp_checks.is_same_experiment(data_exp_fs_manager, exp_config)
+        if same_exp_result is not None:
+            match same_exp_result:
+                case exp_checks.SameExperimentError.DIFFERENT_SYNTAX:
+                    return ErrorStatus.WRONG_EXPERIMENT_CONFIG_SYNTAX
+                case exp_checks.SameExperimentError.NOT_SAME:
+                    return ErrorStatus.DIFFERENT_EXPERIMENT
     else:
         data_exp_fs_manager.exp_dir().mkdir(parents=True, exist_ok=True)
         exp_config.to_yaml(data_exp_fs_manager.config_yaml())
@@ -207,6 +288,16 @@ def _get_samples_to_run(
     run_stats.add_samples_to_run(len(samples_to_run))
 
     return samples_to_run
+
+
+def _prepare_sample_directories(
+    samples_to_run: Iterable[smp_fs.RowNumberedItem],
+    working_exp_fs_manager: exp_fs.Manager,
+) -> None:
+    """Prepare sample directories."""
+    for run_sample in samples_to_run:
+        sample_fs_manager = working_exp_fs_manager.sample_fs_manager(run_sample.item())
+        sample_fs_manager.sample_dir().mkdir(parents=True, exist_ok=True)
 
 
 def _filter_missing_inputs(
@@ -243,6 +334,9 @@ def _write_experiment_missing_inputs(
     run_stats.samples_with_missing_inputs().extend(
         sample.item().exp_sample_id() for sample in samples_with_missing_inputs
     )
+
+    _LOGGER.error("Samples with missing inputs: %d", len(samples_with_missing_inputs))
+
     with exp_errors.ErrorsTSVWriter.open(
         working_exp_fs_manager.errors_tsv(),
         "w",
@@ -272,7 +366,7 @@ def _create_and_run_sbatch_script(  # noqa: PLR0913
         tool_inputs,
         working_exp_fs_manager,
     )
-    script_path = exp_shell.create_run_script(
+    exp_shell.create_run_script(
         data_exp_fs_manager,
         working_exp_fs_manager,
         checked_inputs_samples_to_run,
@@ -281,8 +375,40 @@ def _create_and_run_sbatch_script(  # noqa: PLR0913
     )
 
     cmd_path = subprocess_lib.command_path(slurm.SBATCH_CMD)
-    cli_line = [cmd_path, script_path]
+    cli_line = [cmd_path, working_exp_fs_manager.sbatch_sh_script()]
     subprocess_lib.run_cmd(cli_line, slurm.SBATCH_CMD)
+
+
+def _wait_all_job_finish(
+    checked_inputs_samples_to_run: list[smp_fs.RowNumberedItem],
+    working_exp_fs_manager: exp_fs.Manager,
+) -> None:
+    """Wait all job finish."""
+    in_running_jobs = checked_inputs_samples_to_run.copy()
+
+    with rich_prog.Progress(console=root_logging.CONSOLE) as progress:
+        slurm_running_task = progress.add_task(
+            "Slurm running",
+            total=len(in_running_jobs),
+        )
+        while in_running_jobs:
+            time.sleep(60)
+            _tmp_running_jobs = []
+            for job in in_running_jobs:
+                sample_fs_manager = working_exp_fs_manager.sample_fs_manager(
+                    job.item(),
+                )
+                if (
+                    smp_status.get_status(sample_fs_manager)
+                    == smp_status.ErrorStatus.NOT_RUN
+                ):
+                    _tmp_running_jobs.append(job)
+
+            progress.update(
+                slurm_running_task,
+                advance=(len(in_running_jobs) - len(_tmp_running_jobs)),
+            )
+            in_running_jobs = _tmp_running_jobs
 
 
 def _write_experiment_errors(
@@ -297,6 +423,8 @@ def _write_experiment_errors(
         )
         if smp_status.get_status(sample_fs_manager) == smp_status.ErrorStatus.ERROR:
             run_stats.samples_with_errors().append(run_sample.item().exp_sample_id())
+
+    _LOGGER.error("Samples with errors: %d", len(run_stats.samples_with_errors()))
 
     with exp_errors.ErrorsTSVWriter.open(
         working_exp_fs_manager.errors_tsv(),
@@ -358,8 +486,11 @@ def _move_work_to_data(
     #
     # Move experiment scripts
     #
-    for work_script_file in working_exp_fs_manager.scripts_dir().iterdir():
-        shutil.move(work_script_file, data_exp_fs_manager.scripts_dir())
+    for sh_script in (
+        working_exp_fs_manager.sbatch_sh_script(),
+        working_exp_fs_manager.command_sh_script(),
+    ):
+        shutil.move(sh_script, data_exp_fs_manager.scripts_dir())
     #
     # Move experiment errors
     #