+++ src/pbfbench/topics/seeds/platon/shell.py
@@ -1,6 +1,8 @@
 """Platon Bash script logics."""
 
 from collections.abc import Iterator
+from pathlib import Path
+from typing import final
 
 import pbfbench.abc.tool.shell as abc_tool_shell
 import pbfbench.samples.shell as smp_sh
@@ -9,6 +11,7 @@ import pbfbench.topics.assembly.results.items as asm_results
 import pbfbench.topics.seeds.platon.config as platon_cfg
 
 
+@final
 class GenomeInputLinesBuilder(
     abc_tool_shell.ArgBashLinesBuilder[asm_results.FastaGZ],
 ):
@@ -17,14 +20,22 @@ class GenomeInputLinesBuilder(
     FASTA_GZ_VAR = sh.Variable("fasta_gz")
     FASTA_VAR = sh.Variable("fasta")
 
+    def __fasta_tmp_file(self) -> Path:
+        """Return a tmp FASTA path with sample name is a sh variable."""
+        return self._sample_sh_var_fs_manager.sample_dir() / "tmp_assembly.fasta"
+
     def init_lines(self) -> Iterator[str]:
         """Get shell input init lines."""
-        fasta_gz_path = self.__tool_data_result.fasta_gz(
+        fasta_gz_path = self._tool_data_result.fasta_gz(
             smp_sh.SpeSmpIDLinesBuilder.SPE_SMP_ID_VAR.eval(),
         )
         yield self.FASTA_GZ_VAR.set(sh.path_to_str(fasta_gz_path))
-        yield f"gunzip -k {sh.path_to_str(self.FASTA_GZ_VAR.eval())}"
-        yield self.FASTA_VAR.set(sh.path_to_str(f"${{{self.FASTA_GZ_VAR.name()}%.gz}}"))
+        yield self.FASTA_VAR.set(sh.path_to_str(self.__fasta_tmp_file()))
+        yield (
+            "gunzip -k -c"
+            f" {sh.path_to_str(self.FASTA_GZ_VAR.eval())}"
+            f"> {sh.path_to_str(self.FASTA_VAR.eval())}"
+        )
 
     def argument(self) -> str:
         """Get shell input param lines."""
@@ -35,18 +46,19 @@ class GenomeInputLinesBuilder(
         yield f"rm -f {sh.path_to_str(self.FASTA_VAR.eval())}"
 
 
+@final
 class Commands(abc_tool_shell.Commands[platon_cfg.Names, platon_cfg.Options]):
     """Platon commands."""
 
     def core_commands(self) -> Iterator[str]:
         """Iterate over the tool commands."""
         outdir = smp_sh.sample_sh_var_fs_manager(
-            self.__working_exp_fs_manager,
+            self._working_exp_fs_manager,
         ).sample_dir()
         # REFACTOR how to generalize according to different options types?
         yield (
             "platon "
-            + " ".join(self.__options)
+            + " ".join(self._options)
             + f' --output "{outdir}"'
             + " "
             + self.argument(platon_cfg.Names.GENOME)